backend/pyproject.toml
[tool.poetry]
name = "instruct-ai"
version = "0.1.0"
description = ""
authors = ["Vyoman Jain <75789232+VyoJ@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.115.2"
llama-index = "^0.11.19"
llama-index-embeddings-jinaai = "^0.3.1"
python-dotenv = "^1.0.1"
uvicorn = "^0.32.0"
llama-index-llms-openai-like = "^0.2.0"
groq = "^0.11.0"
llama-index-llms-groq = "^0.2.0"
openai = "^1.52.0"
python-multipart = "^0.0.12"
llama-index-vector-stores-pinecone = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
